+++ polog/config.py
@@ -1,5 +1,3 @@
-from polog.errors import DatabaseIsAlreadyDeterminedError
-from polog.connector import Connector
 from polog.base_settings import BaseSettings
 from polog.levels import Levels
 
@@ -14,26 +12,12 @@ class config(object):
         'level': (int, str),
         'errors_level': (int, str),
         'original_exceptions': (bool, ),
+        'delay_before_exit': (float, int, )
     }
     convert_values = {
         'level': Levels.get,
         'errors_level': Levels.get,
     }
-
-    @classmethod
-    def db(cls, **kwargs):
-        """
-        Логгер Polog использует Pony ORM. В данный метод передаются аргументы так же, как в метод db.bind() самой Pony.
-        См.: https://docs.ponyorm.org/database.html
-
-        Скажем, базу данных sqlite можно инициализировать, вызвав данный метод как-то так:
-        Config.db(provider='sqlite', filename=os.path.join(os.getcwd(), 'logs.db'), create_db=True)
-        (именно так это и будет сделано по умолчанию, если этот метод  не будет вызван.)
-        """
-        if hasattr(cls, 'db_determined'):
-            raise DatabaseIsAlreadyDeterminedError('You have already defined a database for logging. Forgot?')
-        connect = Connector(**kwargs)
-        setattr(cls, 'db_determined', True)
 
     @staticmethod
     def set(**kwargs):
@@ -77,7 +61,7 @@ class config(object):
             Levels.set(key, value)
 
     @staticmethod
-    def add_handler(*args):
+    def add_handlers(*args):
         """
         Добавляем дополнительные обработчики для логов.
         """
