--- polog/db.py
@@ -1,4 +1,0 @@
-from pony.orm import Database
-
-
-db = Database()
