+++ polog/handlers/pony/select.py
@@ -1,6 +1,6 @@
 from pony.orm import db_session, count
 from polog.model import Log
-from polog.utils.select_items import select_items
+from polog.handlers.pony.select_items import select_items
 
 
 logs = Log
